+++ R/AAAparameter.R
@@ -1,6 +1,6 @@
 #######################################################################
 # Code to check parameter/control objects
-# Copyrigth (C) 2011 Michael Hahsler
+# Copyrigth (C) 2021 Michael Hahsler
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -25,6 +25,17 @@
 }
 
 .get_parameters <- function(parameter, defaults) {
+  if(!is.null(parameter[["help"]])){
+    cat("Available control parameters (with default values):\n")
+    cat(rbind(names(defaults)," = ", gsub("\n"," ",as.character(defaults))), 
+      sep=c("\t"," ","\n"))
+    
+    # stop quietly
+    opt <- options(show.error.messages = F)
+    on.exit(options(opt))
+    stop()
+  }
+  
   defaults <- as.list(defaults)
   parameter <- as.list(parameter)