+++ R/plot.R
@@ -1,6 +1,6 @@
 #######################################################################
 # arulesViz - Visualizing Association Rules and Frequent Itemsets
-# Copyrigth (C) 2011 Michael Hahsler and Sudheer Chelluboina
+# Copyrigth (C) 2021 Michael Hahsler
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -29,14 +29,20 @@ plot.rules <- function(x, method = NULL,
     "doubledecker",
     "graph",
     "paracoord",
-    "scatterplot",	## standard
+    "scatterplot",	## default
     "grouped matrix",
     "two-key plot",
     "matrix3D"
   )
-  
+ 
+  if(!is.null(method) && method == "help") {
+    message("Available methods for plotting rules are:\n", paste0(methods, collapse = ", "))
+    return(invisible(methods))  
+  }
+   
   if(length(x)<1) stop("x contains 0 rules!")
   
+  ### default is a scatter plot
   if(is.null(method)) methodNr <- 6
   else methodNr <- pmatch(tolower(method), tolower(methods))
   if(is.na(methodNr)) stop (paste("Unknown method:",sQuote(method), "\nAvailable methods:", paste(sQuote(methods), collapse = ", ")))
@@ -53,7 +59,8 @@ plot.rules <- function(x, method = NULL,
   ## NA means no shading
   if(!all(is.na(shading))) {
     sid <- pmatch(shading, colnames(quality(x)))
-    if(any(is.na(sid))) stop("Shading measure not available in rule set: ",       paste(sQuote(shading[is.na(sid)]), collapse = ", "))
+    if(any(is.na(sid))) stop("Shading measure not available in rule set: ",       
+      paste(sQuote(shading[is.na(sid)]), collapse = ", "))
     shading <- colnames(quality(x))[sid]
   }
    
@@ -65,43 +72,50 @@ plot.rules <- function(x, method = NULL,
     interactive <- FALSE
   }
    
+  ### if control is character then I think it is "help"
+  if(is.character(control)) control <- list(help = TRUE)
   if(is.null(control$engine)) control$engine <- engine
   
-  
-  
   ## work horses
-  if (methodNr == 1) matrix_arules(x, measure = shading, control,...)
-  else if (methodNr == 2) doubledecker_arules(x, measure = measure, 
+  if (methodNr == 1) matrixplot(x, measure = shading, control,...)
+  
+  else if (methodNr == 2) doubledeckerplot(x, measure = measure, 
     data = data, c(control, list(type="mosaic")), ...)
-  else if (methodNr == 3) doubledecker_arules(x, measure = measure, 
-    data = data, control, ...)
-  else if (methodNr == 4) graph_arules(x, measure = measure, 
+  
+  else if (methodNr == 3) doubledeckerplot(x, measure = measure, 
+    data = data, c(control, list(type="doubledecker")), ...)
+  
+  else if (methodNr == 4) graphplot(x, measure = measure, 
     shading = shading, control, ...)
-  else if (methodNr == 5) paracoord_arules(x, measure = measure, 
-    shading = shading, control = control,...)
+  
+  else if (methodNr == 5) paracoord_rules(x, measure = measure, 
+    shading = shading, control = control, ...)
+  
   else if (methodNr == 6) {
     if(length(measure)<2) measure[2] <- "confidence"
-    scatterplot_arules(x, measure = measure, 
-      shading = shading, control, ...)
+    scatterplot(x, measure = measure, shading = shading, control, ...)
   }
-  else if (methodNr == 7) grouped_matrix_arules(x, measure= measure, 
+  
+  else if (methodNr == 7) grouped_matrix_plot(x, measure= measure, 
     shading = shading, control=control, ...)
+  
   else if (methodNr == 8) { 
-    if(is.null(control$col)) control$col <- rainbow(max(size(x))-1L)
-    scatterplot_arules(x, 
+    if(is.null(control$col)) control$col <- grDevices::rainbow(max(size(x))-1L)
+    scatterplot(x, 
     measure = c("support", "confidence"), shading = "order", 
     control, ...)
   }
+  
   else if (methodNr == 9) {
     warning("method 'matrix3D' is deprecated use method 'matrix' with engine '3d'")
     control$engine <- "3d"
-    matrix_arules(x, measure = shading, control = control, ...)
+    matrixplot(x, measure = shading, control = control, ...)
   }
 }
 
 plot.itemsets <- function(x, method = NULL, 
   measure = "support", shading = NA, 
-  interactive=NULL, engine = "default",
+  interactive = NULL, engine = "default",
   data = NULL, control = NULL, ...) {
   
   ## methods
@@ -110,6 +124,11 @@ plot.itemsets <- function(x, method = NULL,
     "paracoord",
     "scatterplot"
   )
+  
+  if(!is.null(method) && method == "help") {
+    message("Available methods for plotting itemsets are:\n", paste0(methods, collapse = ", "))
+    return(invisible(methods))  
+  }
   
   ## add interactive and engine
   if(is.null(control$engine)) control$engine <- engine
@@ -131,14 +150,15 @@ plot.itemsets <- function(x, method = NULL,
   
   
   ## work horses
-  if (methodNr == 1) graph_arules(x, measure = measure,
+  if (methodNr == 1) graphplot(x, measure = measure, shading = shading,
     control= control, ...)
-  else if (methodNr == 2) paracoord_items(x, measure = measure,
+  
+  else if (methodNr == 2) paracoord_items(x, measure = measure, shading = shading,
     control= control, ...)
+  
   else if (methodNr == 3) {
     if(length(measure)<2) measure[2] <- "order"
-    scatterplot_arules(x, measure = measure, 
-      shading = shading, control, ...)
+    scatterplot(x, measure = measure, shading = shading, control, ...)
   }
   
 }