+++ NEWS.md
@@ -1,3 +1,18 @@
+# arulesViz 1.5-0 (05/21/2021)
+
+## New features
+* ggplot2 is now the default engine for most plots. 
+* grouped matrix visualization has now engine ggplot and htmlwidget.
+* conversion from associations to igraph are now exposed.
+* conversion from rules to matrix and groupedMatrix are now exposed.
+* "help" can now be specified for plot for method, engine, and control. 
+* ruleExplorer lost parameter and gained sidebarWidth and chartHeight.
+
+## Bug Fixes
+* Fixed rounding bug in inspectDT.
+* Fixed jitter for scatter plot using engine plotly.
+* Fixed reordering for grouped matrix (engine plotly).
+
 # arulesViz 1.4-0 (03/05/2021)
 
 ## New features
@@ -10,7 +25,7 @@
 
 ## Bug Fixes
 * fixed bug that prevented shading = NA in plotting method graph.
-* fixed some control parameter handling and added more example for plot method graph.
+* fixed some control parameter handling and added more examples for plot method graph.
 * fixed additional item shown in LHS for grouped matrix plot (reported by Tasos Fasoulis).
 
 # arulesViz 1.3-3 (05/20/2019)